Extract admin pubs list
